+++ src/display/side_by_side.rs
@@ -37,6 +37,7 @@ fn format_missing_line_num(
     prev_num: LineNumber,
     source_dims: &SourceDimensions,
     side: Side,
+    is_continuation: bool,
     use_color: bool,
 ) -> String {
     let column_width = match side {
@@ -45,8 +46,8 @@ fn format_missing_line_num(
     };
 
     let after_end = match side {
-        Side::Left => prev_num >= source_dims.lhs_max_line,
-        Side::Right => prev_num >= source_dims.rhs_max_line,
+        Side::Left => prev_num >= source_dims.lhs_max_line_in_file,
+        Side::Right => prev_num >= source_dims.rhs_max_line_in_file,
     };
 
     let mut style = Style::new();
@@ -54,10 +55,23 @@ fn format_missing_line_num(
         style = style.dimmed();
     }
 
+    let c = if is_continuation {
+        // Always show dots when this line is too long so we had to
+        // split it over several lines when displaying.
+        "."
+    } else if after_end {
+        // The file on this side has fewer lines, and we've already
+        // shown the last ones.
+        " "
+    } else {
+        // There are more lines in this file.
+        "."
+    };
+
     let num_digits = prev_num.display().len();
     format!(
         "{:>width$} ",
-        (if after_end { " " } else { "." }).repeat(num_digits),
+        c.repeat(num_digits),
         width = column_width - 1
     )
     .style(style)
@@ -127,6 +141,7 @@ fn display_line_nums(
             prev_lhs_line_num.unwrap_or_else(|| 1.into()),
             source_dims,
             Side::Left,
+            false,
             display_options.use_color,
         ),
     };
@@ -139,6 +154,7 @@ fn display_line_nums(
             prev_rhs_line_num.unwrap_or_else(|| 1.into()),
             source_dims,
             Side::Right,
+            false,
             display_options.use_color,
         ),
     };
@@ -158,32 +174,21 @@ struct SourceDimensions {
     /// The number of characters required to display line numbers on
     /// the RHS.
     rhs_line_nums_width: usize,
-    /// The highest line number in the LHS source.
-    lhs_max_line: LineNumber,
-    /// The highest line number in the RHS source.
-    rhs_max_line: LineNumber,
+    lhs_max_line_in_file: LineNumber,
+    rhs_max_line_in_file: LineNumber,
 }
 
 impl SourceDimensions {
     fn new(
         terminal_width: usize,
-        line_nums: &[(Option<LineNumber>, Option<LineNumber>)],
+        lhs_max_line_visible: LineNumber,
+        rhs_max_line_visible: LineNumber,
+        lhs_max_line_in_file: LineNumber,
+        rhs_max_line_in_file: LineNumber,
         content_max_width: usize,
     ) -> Self {
-        let mut lhs_max_line: LineNumber = 1.into();
-        let mut rhs_max_line: LineNumber = 1.into();
-
-        for (lhs_line_num, rhs_line_num) in line_nums {
-            if let Some(lhs_line_num) = lhs_line_num {
-                lhs_max_line = max(lhs_max_line, *lhs_line_num);
-            }
-            if let Some(rhs_line_num) = rhs_line_num {
-                rhs_max_line = max(rhs_max_line, *rhs_line_num);
-            }
-        }
-
-        let lhs_line_nums_width = format_line_num(lhs_max_line).len();
-        let rhs_line_nums_width = format_line_num(rhs_max_line).len();
+        let lhs_line_nums_width = format_line_num(lhs_max_line_visible).len();
+        let rhs_line_nums_width = format_line_num(rhs_max_line_visible).len();
 
         // If the file lines are extremely short, treat them as if
         // they have a line of 25 characters.
@@ -236,8 +241,8 @@ impl SourceDimensions {
             content_display_width: content_width,
             lhs_line_nums_width,
             rhs_line_nums_width,
-            lhs_max_line,
-            rhs_max_line,
+            lhs_max_line_in_file,
+            rhs_max_line_in_file,
         }
     }
 }
@@ -328,6 +333,88 @@ fn highlight_as_novel(
     false
 }
 
+/// Find the longest line in `lhs_src` and `rhs_src` that will be
+/// displayed.
+fn displayed_content_max_len_in_bytes(
+    lhs_src: &str,
+    rhs_src: &str,
+    hunks: &[Hunk],
+    num_context_lines: u32,
+) -> usize {
+    let mut lhs_displayed_lines: DftHashSet<usize> = DftHashSet::default();
+    let mut rhs_displayed_lines: DftHashSet<usize> = DftHashSet::default();
+
+    for hunk in hunks {
+        let mut min_lhs_line: Option<LineNumber> = None;
+        let mut max_lhs_line: Option<LineNumber> = None;
+        let mut min_rhs_line: Option<LineNumber> = None;
+        let mut max_rhs_line: Option<LineNumber> = None;
+
+        for (lhs_line, rhs_line) in &hunk.lines {
+            if let Some(lhs_line) = lhs_line {
+                if let Some(current_min) = min_lhs_line {
+                    min_lhs_line = Some(min(current_min, *lhs_line));
+                } else {
+                    min_lhs_line = Some(*lhs_line);
+                }
+
+                if let Some(current_max) = max_lhs_line {
+                    max_lhs_line = Some(max(current_max, *lhs_line));
+                } else {
+                    max_lhs_line = Some(*lhs_line);
+                }
+            }
+
+            if let Some(rhs_line) = rhs_line {
+                if let Some(current_min) = min_rhs_line {
+                    min_rhs_line = Some(min(current_min, *rhs_line));
+                } else {
+                    min_rhs_line = Some(*rhs_line);
+                }
+
+                if let Some(current_max) = max_rhs_line {
+                    max_rhs_line = Some(max(current_max, *rhs_line));
+                } else {
+                    max_rhs_line = Some(*rhs_line);
+                }
+            }
+
+            if let (Some(min_lhs_line), Some(max_lhs_line)) = (min_lhs_line, max_lhs_line) {
+                let min_lhs_plus_padding =
+                    max(0, min_lhs_line.0 as isize - num_context_lines as isize) as usize;
+                let max_lhs_plus_padding = max_lhs_line.0 as usize + num_context_lines as usize;
+                for lhs_line_num in min_lhs_plus_padding..=max_lhs_plus_padding {
+                    lhs_displayed_lines.insert(lhs_line_num);
+                }
+            }
+
+            if let (Some(min_rhs_line), Some(max_rhs_line)) = (min_rhs_line, max_rhs_line) {
+                let min_rhs_plus_padding =
+                    max(0, min_rhs_line.0 as isize - num_context_lines as isize) as usize;
+                let max_rhs_plus_padding = max_rhs_line.0 as usize + num_context_lines as usize;
+                for rhs_line_num in min_rhs_plus_padding..=max_rhs_plus_padding {
+                    rhs_displayed_lines.insert(rhs_line_num);
+                }
+            }
+        }
+    }
+
+    let mut content_max_width: usize = 0;
+
+    for (lhs_i, lhs_line) in lhs_src.lines().enumerate() {
+        if lhs_displayed_lines.contains(&lhs_i) {
+            content_max_width = max(content_max_width, lhs_line.len());
+        }
+    }
+    for (rhs_i, rhs_line) in rhs_src.lines().enumerate() {
+        if rhs_displayed_lines.contains(&rhs_i) {
+            content_max_width = max(content_max_width, rhs_line.len());
+        }
+    }
+
+    content_max_width
+}
+
 pub(crate) fn print(
     hunks: &[Hunk],
     display_options: &DisplayOptions,
@@ -339,13 +426,12 @@ pub(crate) fn print(
     lhs_mps: &[MatchedPos],
     rhs_mps: &[MatchedPos],
 ) {
-    let mut content_max_width: usize = 0;
-    for line in lhs_src.lines() {
-        content_max_width = max(content_max_width, line.len());
-    }
-    for line in rhs_src.lines() {
-        content_max_width = max(content_max_width, line.len());
-    }
+    let content_max_width = displayed_content_max_len_in_bytes(
+        lhs_src,
+        rhs_src,
+        hunks,
+        display_options.num_context_lines,
+    );
 
     let (lhs_colored_lines, rhs_colored_lines) = if display_options.use_color {
         (
@@ -458,6 +544,52 @@ pub(crate) fn print(
     let matched_lines = all_matched_lines_filled(lhs_mps, rhs_mps, &lhs_lines, &rhs_lines);
     let mut matched_lines_to_print = &matched_lines[..];
 
+    let mut lhs_max_visible_line = 1.into();
+    let mut rhs_max_visible_line = 1.into();
+
+    if let Some(hunk) = hunks.last() {
+        let (start_i, end_i) = matched_lines_indexes_for_hunk(
+            matched_lines_to_print,
+            hunk,
+            display_options.num_context_lines as usize,
+        );
+        let aligned_lines = &matched_lines_to_print[start_i..end_i];
+
+        for (lhs_line_num, rhs_line_num) in aligned_lines.iter().rev() {
+            if let Some(lhs_line_num) = *lhs_line_num {
+                lhs_max_visible_line = max(lhs_max_visible_line, lhs_line_num);
+            }
+            if let Some(rhs_line_num) = *rhs_line_num {
+                rhs_max_visible_line = max(rhs_max_visible_line, rhs_line_num);
+            }
+
+            if lhs_max_visible_line > 1.into() && rhs_max_visible_line > 1.into() {
+                break;
+            }
+        }
+    }
+
+    let lhs_max_line_in_file = LineNumber(lhs_lines.len().saturating_sub(1) as u32);
+    let rhs_max_line_in_file = LineNumber(rhs_lines.len().saturating_sub(1) as u32);
+
+    lhs_max_visible_line = LineNumber(min(
+        lhs_max_visible_line.0 + display_options.num_context_lines,
+        lhs_max_line_in_file.0,
+    ));
+    rhs_max_visible_line = LineNumber(min(
+        rhs_max_visible_line.0 + display_options.num_context_lines,
+        rhs_max_line_in_file.0,
+    ));
+
+    let source_dims = SourceDimensions::new(
+        display_options.terminal_width,
+        lhs_max_visible_line,
+        rhs_max_visible_line,
+        lhs_max_line_in_file,
+        rhs_max_line_in_file,
+        content_max_width,
+    );
+
     for (i, hunk) in hunks.iter().enumerate() {
         println!(
             "{}",
@@ -488,11 +620,6 @@ pub(crate) fn print(
         let no_rhs_changes = hunk.novel_rhs.is_empty();
         let same_lines = aligned_lines.iter().all(|(l, r)| l == r);
 
-        let source_dims = SourceDimensions::new(
-            display_options.terminal_width,
-            aligned_lines,
-            content_max_width,
-        );
         for (lhs_line_num, rhs_line_num) in aligned_lines {
             let lhs_line_novel = highlight_as_novel(
                 *lhs_line_num,
@@ -596,6 +723,7 @@ pub(crate) fn print(
                                 .unwrap_or_else(|| prev_lhs_line_num.unwrap_or_else(|| 10.into())),
                             &source_dims,
                             Side::Left,
+                            true,
                             display_options.use_color,
                         );
                         if let Some(line_num) = lhs_line_num {
@@ -616,6 +744,7 @@ pub(crate) fn print(
                                 .unwrap_or_else(|| prev_rhs_line_num.unwrap_or_else(|| 10.into())),
                             &source_dims,
                             Side::Right,
+                            true,
                             display_options.use_color,
                         );
                         if let Some(line_num) = rhs_line_num {
@@ -657,8 +786,14 @@ mod tests {
 
     #[test]
     fn test_width_calculations() {
-        let line_nums = [(Some(1.into()), Some(10.into()))];
-        let source_dims = SourceDimensions::new(DEFAULT_TERMINAL_WIDTH, &line_nums, 9999);
+        let source_dims = SourceDimensions::new(
+            DEFAULT_TERMINAL_WIDTH,
+            1.into(),
+            10.into(),
+            1.into(),
+            10.into(),
+            9999,
+        );
 
         assert_eq!(source_dims.lhs_line_nums_width, 2);
         assert_eq!(source_dims.rhs_line_nums_width, 3);
@@ -668,40 +803,33 @@ mod tests {
     fn test_format_missing_line_num() {
         let source_dims = SourceDimensions::new(
             DEFAULT_TERMINAL_WIDTH,
-            &[
-                (Some(0.into()), Some(0.into())),
-                (Some(1.into()), Some(1.into())),
-            ],
+            1.into(),
+            1.into(),
+            1.into(),
+            1.into(),
             9999,
         );
 
         assert_eq!(
-            format_missing_line_num(0.into(), &source_dims, Side::Left, true),
+            format_missing_line_num(0.into(), &source_dims, Side::Left, false, true),
             ". ".dimmed().to_string()
         );
         assert_eq!(
-            format_missing_line_num(0.into(), &source_dims, Side::Left, false),
+            format_missing_line_num(0.into(), &source_dims, Side::Left, false, false),
             ". ".to_owned()
         );
     }
 
     #[test]
     fn test_format_missing_line_num_at_end() {
-        let source_dims = SourceDimensions::new(
-            80,
-            &[
-                (Some(0.into()), Some(0.into())),
-                (Some(1.into()), Some(1.into())),
-            ],
-            9999,
-        );
+        let source_dims = SourceDimensions::new(80, 1.into(), 1.into(), 1.into(), 1.into(), 9999);
 
         assert_eq!(
-            format_missing_line_num(1.into(), &source_dims, Side::Left, true),
+            format_missing_line_num(1.into(), &source_dims, Side::Left, false, true),
             "  ".dimmed().to_string()
         );
         assert_eq!(
-            format_missing_line_num(1.into(), &source_dims, Side::Left, false),
+            format_missing_line_num(1.into(), &source_dims, Side::Left, false, false),
             "  ".to_owned()
         );
     }